added mailing system

+++ api/usercart/views.py
@@ -26,6 +26,10 @@ def validate_user_session(id, token):
 @csrf_exempt
 def add(request, user_id,token):
     # return JsonResponse({'msg': 'HI'})
+
+
+
+    print("called")
     if not validate_user_session(user_id, token):
         return JsonResponse({'error': 'Please re-login', 'code': '1'})
     if request.method == "POST":
@@ -46,19 +50,41 @@ def add(request, user_id,token):
         except Product.DoesNotExist:
             return JsonResponse({'error': 'Product does not exist'})
             # i have to check if product is already present in the given usercart or not
+        print("qty=",quantity)
+        print("selectedProductColor=",selectedProductColor)
+        print("selectedProductSize=",selectedProductSize)
         try:
-            previouscart=Usercart.objects.get(user=user_id,product=int(product_id))
-            print("Previouscart",previouscart)
-            newquantity=previouscart.quantity+1
-            previouscart.delete()
-            previouscart = Usercart(user=user,product=product,selectedProductColor=selectedProductColor,selectedProductSize=selectedProductSize,quantity=newquantity)
+            previouscart=Usercart.objects.get(user=user_id,product=int(product_id),selectedProductSize=selectedProductSize)
+            previouscart.quantity=previouscart.quantity+int(quantity)
             previouscart.save()
             return JsonResponse({'success': True, 'error': False,'id':previouscart.pk})
+            
+            # print("end")
         except Usercart.DoesNotExist:
             cart = Usercart(user=user,product=product,selectedProductColor=selectedProductColor,selectedProductSize=selectedProductSize,quantity=int(quantity))
             cart.save()
             return JsonResponse({'success': True, 'error': False,'id':cart.pk})
-        
+def deleteall(request,user_id):
+    try:
+        cart=Usercart.objects.all().filter(user=user_id)
+        print(cart)
+        cart.delete()
+        return JsonResponse({'success': True, 'error': False})
+    except Usercart.DoesNotExist:
+        return JsonResponse({'error': 'Cart does not exist'})
+
             
+def decrease(request, user_id,cart_id):
     
-    return JsonResponse({'success': True, 'error': False})
+    print(user_id,cart_id)
+    try:
+        previouscart=Usercart.objects.get(id=cart_id)
+        previouscart.quantity=previouscart.quantity-1
+        if previouscart.quantity==0:
+            previouscart.delete()
+        else:
+            previouscart.save()
+        # print("end")
+        return JsonResponse({'success': True, 'error': False,'id':previouscart.pk})
+    except Usercart.DoesNotExist:
+        return JsonResponse({'success': False, 'error': True})
